Config is now mutable, added sub-commands /config (config listing and editing) and /model (model setting, listing, refreshing details). CLI used for smart config field setting. UX has more dynamic flow, where model and API can change mid-conversation

+++ solveig/interface/cli/interface.py
@@ -231,8 +231,8 @@ class TerminalInterface(SolveigInterface):
         url: str | None = None,
         path: str | PathLike | None = None,
         max_context: int | None = None,
-        input_price: int | None = None,
-        output_price: int | None = None,
+        input_price: float | None = None,
+        output_price: float | None = None,
     ) -> None:
         """Update stats dashboard with multiple pieces of information."""
         _max_context = (
